Rename method channel wrapper
